+++ example/bridge/base_straight.py
@@ -1,14 +1,14 @@
 import cadquery as cq
 from cqterrain.bridge import BaseStraight
 
-bp_ramp = BaseStraight()
+bp_straight = BaseStraight()
 
-length:float = 75*2
-width:float = 75*2
-height:float = 50
-bp_ramp.make()
+bp_straight.length = 75*2
+bp_straight.width = 75*2
+bp_straight.height = 50
+bp_straight.make()
 
-ex_straight = bp_ramp.build()
+ex_straight = bp_straight.build()
 
-#show_object(ex_ramp)
+#show_object(ex_straight)
 cq.exporters.export(ex_straight,'stl/bridge_straight.stl')