+++ example_runner.py
@@ -2,6 +2,7 @@ print('bridge examples')
 import example.bridge.bridge
 import example.bridge.base_ramp
 import example.bridge.base_straight
+import example.bridge.tile_straight
 
 print('building examples')
 import example.building.building