+++ src/cqterrain/bridge/__init__.py
@@ -1,3 +1,5 @@
 from .BaseStraight import BaseStraight
 from .BaseRamp import BaseRamp
 from .Bridge import Bridge
+
+from .TileStraight import TileStraight