+++ changes.md
@@ -1,5 +1,11 @@
 ## Main wip
 
+## 1.1.1
+* Added bridge TileStraight
+  * Added example
+  * Added Documentation
+* Fixed bridge BaseStraight example
+
 ## 1.1.0
 * Added bridge module
   * Added BaseRamp